Rename orphanrium to disorder.
Fixes #17.
