+++ disorder-aeson/test/Test/Disorder/Aeson.hs
@@ -1,0 +1,15 @@
+{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE OverloadedStrings #-}
+module Test.Disorder.Aeson where
+
+import           Disorder.Aeson
+
+import           Test.QuickCheck
+
+prop_jsonProp :: Int -> Property
+prop_jsonProp =
+  jsonProp
+
+return []
+tests :: IO Bool
+tests = $quickCheckAll
