+++ disorder-core/disorder-core.cabal
@@ -1,0 +1,65 @@
+name:                  disorder-core
+version:               0.0.1
+license:               AllRightsReserved
+author:                Ambiata <info@ambiata.com>
+maintainer:            Ambiata <info@ambiata.com>
+copyright:             (c) 2015 Ambiata
+synopsis:              disorder-core
+category:              System
+cabal-version:         >= 1.8
+build-type:            Simple
+description:           disorder-core.
+
+library
+  build-depends:
+                       base                            >= 3          && < 5
+                     , directory                       == 1.2.*
+                     , text                            >= 1.1        && < 1.3
+                     , process                         == 1.2.*
+                     , QuickCheck                      == 2.7.*
+
+  ghc-options:
+                       -Wall
+
+  hs-source-dirs:
+                       src
+
+
+  exposed-modules:
+                       Paths_disorder_core
+                       Disorder.Core
+                       Disorder.Core.IO
+                       Disorder.Core.Main
+                       Disorder.Core.Property
+                       Disorder.Core.UniquePair
+
+test-suite test
+  type:                exitcode-stdio-1.0
+
+  main-is:             test.hs
+
+  ghc-options:         -Wall -threaded -O2
+
+  hs-source-dirs:
+                       test
+
+  build-depends:
+                       base                            >= 3          && < 5
+                     , disorder-core
+                     , QuickCheck
+                     , text
+                     , quickcheck-instances            == 0.3.*
+
+test-suite test-cli
+  type:                exitcode-stdio-1.0
+
+  main-is:             test-cli.hs
+
+  ghc-options:         -Wall -threaded -O2
+
+  hs-source-dirs:
+                       test
+
+  build-depends:
+                       base                            >= 3          && < 5
+                     , disorder-core
