+++ disorder-core/src/Disorder/Core.hs
@@ -1,0 +1,6 @@
+module Disorder.Core (
+        module OC
+    ) where
+
+import Disorder.Core.Property as OC
+import Disorder.Core.UniquePair as OC
