+++ disorder-core/src/Disorder/Core/IO.hs
@@ -1,0 +1,10 @@
+module Disorder.Core.IO (
+    testIO
+  ) where
+
+import           Test.QuickCheck
+import           Test.QuickCheck.Monadic
+
+
+testIO :: Testable a => IO a -> Property
+testIO = monadicIO . (=<<) stop . run
