+++ disorder-core/test/test-cli.hs
@@ -1,0 +1,5 @@
+import           Disorder.Core.Main
+
+main :: IO ()
+main =
+  disorderCliMain ["ACK"]
