+++ disorder-core/test/test.hs
@@ -1,0 +1,9 @@
+import           Disorder.Core.Main
+import qualified Test.Disorder.Core.Property
+import qualified Test.Disorder.Core.UniquePair
+
+main :: IO ()
+main = disorderMain [
+    Test.Disorder.Core.Property.tests
+  , Test.Disorder.Core.UniquePair.tests
+  ]
