+++ disorder-lens/disorder-lens.cabal
@@ -1,0 +1,44 @@
+name:                  disorder-lens
+version:               0.0.1
+license:               AllRightsReserved
+author:                Ambiata <info@ambiata.com>
+maintainer:            Ambiata <info@ambiata.com>
+copyright:             (c) 2015 Ambiata
+synopsis:              disorder-lens
+category:              System
+cabal-version:         >= 1.8
+build-type:            Simple
+description:           disorder-lens.
+
+library
+  build-depends:
+                       base                            >= 3          && < 5
+                     , lens                            >= 4.6        && < 4.9
+                     , QuickCheck                      == 2.7.*
+
+  ghc-options:
+                       -Wall
+
+  hs-source-dirs:
+                       src
+
+
+  exposed-modules:
+                       Paths_disorder_lens
+                       Disorder.Lens
+
+test-suite test
+  type:                exitcode-stdio-1.0
+
+  main-is:             test.hs
+
+  ghc-options:         -Wall -threaded -O2
+
+  hs-source-dirs:
+                       test
+
+  build-depends:
+                       base                            >= 3          && < 5
+                     , disorder-lens
+                     , lens
+                     , QuickCheck
