--- orphanarium-aeson/orphanarium-aeson.cabal
@@ -1,47 +1,0 @@
-name:                  orphanarium-aeson
-version:               0.0.1
-license:               AllRightsReserved
-author:                Ambiata <info@ambiata.com>
-maintainer:            Ambiata <info@ambiata.com>
-copyright:             (c) 2015 Ambiata
-synopsis:              orphanarium-aeson
-category:              System
-cabal-version:         >= 1.8
-build-type:            Simple
-description:           orphanarium-aeson.
-
-library
-  build-depends:
-                       base                            >= 3          && < 5
-                     , aeson                           == 0.8.*
-                     , QuickCheck                      == 2.7.*
-                     , time
-
-  ghc-options:
-                       -Wall
-
-  hs-source-dirs:
-                       src
-
-
-  exposed-modules:
-                       Paths_orphanarium_aeson
-                       Orphanarium.Aeson
-                       Orphanarium.Aeson.Time
-
-test-suite test
-  type:                exitcode-stdio-1.0
-
-  main-is:             test.hs
-
-  ghc-options:         -Wall -threaded -O2
-
-  hs-source-dirs:
-                       test
-
-  build-depends:
-                       base                            >= 3          && < 5
-                     , orphanarium-aeson
-                     , aeson
-                     , QuickCheck
-                     , time
