--- orphanarium-core/src/Orphanarium/Core.hs
@@ -1,6 +1,0 @@
-module Orphanarium.Core (
-        module OC
-    ) where
-
-import Orphanarium.Core.Property as OC
-import Orphanarium.Core.UniquePair as OC
