--- orphanarium-core/src/Orphanarium/Core/IO.hs
@@ -1,10 +1,0 @@
-module Orphanarium.Core.IO (
-    testIO
-  ) where
-
-import           Test.QuickCheck
-import           Test.QuickCheck.Monadic
-
-
-testIO :: Testable a => IO a -> Property
-testIO = monadicIO . (=<<) stop . run
