--- orphanarium-core/test/test-cli.hs
@@ -1,5 +1,0 @@
-import           Orphanarium.Core.Main
-
-main :: IO ()
-main =
-  orphanariumCliMain ["ACK"]
