--- orphanarium-corpus/orphanarium-corpus.cabal
@@ -1,44 +1,0 @@
-name:                  orphanarium-corpus
-version:               0.0.1
-license:               AllRightsReserved
-author:                Ambiata <info@ambiata.com>
-maintainer:            Ambiata <info@ambiata.com>
-copyright:             (c) 2015 Ambiata
-synopsis:              orphanarium-corpus
-category:              System
-cabal-version:         >= 1.8
-build-type:            Simple
-description:           orphanarium-corpus.
-
-library
-  build-depends:
-                       base                            >= 3          && < 5
-                     , text                            >= 1.1        && < 1.3
-                     , QuickCheck                      == 2.7.*
-
-  ghc-options:
-                       -Wall
-
-  hs-source-dirs:
-                       src
-
-
-  exposed-modules:
-                       Paths_orphanarium_corpus
-                       Orphanarium.Corpus
-
-test-suite test
-  type:                exitcode-stdio-1.0
-
-  main-is:             test.hs
-
-  ghc-options:         -Wall -threaded -O2
-
-  hs-source-dirs:
-                       test
-
-  build-depends:
-                       base                            >= 3          && < 5
-                     , orphanarium-corpus
-                     , text
-                     , QuickCheck
