--- orphanarium-corpus/test/test.hs
@@ -1,13 +1,0 @@
-import           Control.Monad
-
-import qualified Test.Orphanarium.Corpus
-
-import           System.Exit
-import           System.IO
-
-
-main :: IO ()
-main =
-  hSetBuffering stdout LineBuffering >> mapM id [
-      Test.Orphanarium.Corpus.tests
-    ] >>= \rs -> when (not . all id $ rs) exitFailure
